Try a different strategy when determining whether a map is 'clicky' or not
the main thing we need to determine clickiness for is whether
or not the map should be scaled to fit. someday this will
be even more important when clicky maps scale properly.

the current method is to see if the first map event defined
includes a 'click_activate!' comment. it also checks that there
is at least one 'click_activate!' item that does not have
the "thought bubble" icon.

if the current map is determined to be clicky, only 'click_activate!'
things are allowed to be clicked.