Properly render tile-based sprites (like the ottoman in Milf's Villa upstairs)
Fix Milf's villa upstairs room to not allow you to drag the ottoman
until you've made the hole to sister's room. Annoyingly, this is a bug
in the actual game, it's just unusual in the RPGM version that anyone
would choose to move the ottoman onto the area the hole is going to be
before making the hole.

Add some extra code to hopefully recover saves created by people
who moved the ottoman onto the hole before the hole existed. Which
is probably only me.