Show pictures in map screen rather than relying on renpy.hide / renpy.show
This allows pictures to be transformed appropriately (x/y/size/etc)
while still showing inbetween the map and the clickables on the map.

Requires shifting around the interaction model a little bit so that the
map is 'shown' more often and displayed text can be rendered on top of it.

The plus side is now and menus show on top of the map image, rather
than transporting you to a mysterious void.

Also required ensuring that the rendered image on a GameMapBackground
didn't get pickled. Seems like this always should've been the case,
somehow it only came up here.