Fix file loading to work in Linux
For whatever reason, renpy.loader.load doesn't work for absolute
paths on linux because it strips the leading slash, e.g.
/home/foo becomes home/foo

The fix here is to add config.basedir to the searchpath (technically
adding '', which gets joined to config.basedir) and only give renpy
relative paths to that dir.

That also required changing all the calls to os.path.exist and
os.listdir which took an rpgm_path to create their own absolute paths.
Whee.