Pathfinding algorithm no longer prematurely marks a square as visited
This was causing problems for grid squares that are visible but not from
all direct routes; they would be prematurely marked as visited before the
directional reachability was checked, and then no other paths to the
square would be considered.

It could also be possible to ignore directional reachbility because it's
already known whether a square is absolutely reachable or not, but that
might mean in certain situations the PC is pointed in the wrong direction.

Fixes certain situations like going down the stairs in Milf's Villa, e.g.:

.....XPX.......
.....EED.......
where 'X' is empty space, 'P' is the player,
'E' are other exits, and 'D' is the desired exit