Add a fancier reachability grid cache that gets more hits
The cache will reuse the same computed grid if the player
moves to a new position that is reachable from any old
computed grid for the same set of events and pages.

This allows the cache to maintain distinct sets of
reachability grids for the same map (e.g., a house with
multiple enclosed rooms) but not re-compute the entire
grid every time the player looks at a plant or something
in a room.