Better preserve scroll position on load
Previously, when you loaded, the global 'viewport_xadjustment' and the
mapscreen's viewport.xadjustment would be different objects until
the next time you triggered a full screen update.

So scrolling would appear to work right, but once you zoomed in
the viewport's xadjustment would get replaced with the global one
and the scroll position would not be what you expect.

Now, when loading a save game, the global viewport_xadjustment is
replaced with the one from the displayed map screen if available.

especially helpful for clicky screens in milfs-villa, which
previously would have their clickables way off screen on load