Milf's Villa: even better positioning of clickables on clicky screens
The normal map zoom code ignores blank areas at the bottom/right of the
map, but the clicky screen code needs to consider the entire map so that
it can show clickable events in the right place.

Also: ensure the range of the x/y adjustments are set correctly when
transitioning from clicky to non-clicky maps, otherwise the value set
to scroll the clickies into view might be lost.