Remove moving cars from the map in ics1
They often spawn near the map entrance/exit and block forward progress
if the player happens to spawn there.
This isn't a problem in RPGM because they move away, but in RenPy they'll
just stay there and block progress forever.
There are also streetlights that cause a similar problem, but I'm not
going to fix it because it's an actual RPGM game bug and it's kinda funny.