Only show the last 'rpgmpicture-scene' in ICS1
There are some scenes in ICS1 where pictures are repeatedly shown on top
of eachother. RenPy has to create textures for all these pictures and
starts to creak and slow down after a while, so it's better to only show
the topmost picture. Since all the `scene` pictures in ICS1 seem to be
opaque, this shouldn't cause a problem.