Add a script to export quests from the VX Ace 'Quest_Log.rb' script
Since these scripts are written in ruby, the easiest way is to just
set up a bunch of fake RPGM objects and run the ruby code.

Requires manually passing in the quest ID for the last quest, since
quests don't exist as a big array of data but as a method you pass
a quest ID into.

Needex for ic adventure / the artifact