Special case handling of 'd-pad' style event wait
In certain RPGM games (IA, The Artifact) the action is paused while a
parallel event plays an animation. The author elected to surround the
player with 4 identical events that advance to the next scene, such that
if the player presses the action button or moves in any direction they
will continue.

Since it's pretty obnoxious to have a person playing the RenPy game click
on some random d-pad shaped squares on the screen, the engine attempts to
auto-advance if it thinks it's in this situation.

One problem is that not 100% of these scenes are authored consistently; in
one of them, one of the surrounding events is two spaces away. So that
will look weird.

Also, the method of advancing doesn't successfully hide other background
events while the next scene is showing, so it kind of relies of
'hide_unreachable_events' being set true.