Add the ability to conditionally specify 'curated_walk_locations'
In theory, every reachable square could be a walk location, but that would
bring RenPy to its knees in the current implementation.

This approach allows certain puzzles that require the player to be in a
certain place to be worked around, though it requires game-specific code.

Needed for the top floor of the lab in IA.