Animate certain map sprites
Event pages that are set as 'stepAnime' are supposed to be animated

For the time being I am only animating those events which are also
"object" sprites, meaning they have a filename starting with an
exclamation, like '!Flames'. I'm worried that too many animations
on a big map will make the game even slower than it already is.

The main goal of this for now is to make the stars animate in
My Girlfriend's Amnesia