Allow rollback, but only to the start of a given event chain.
This seems to prevent most of the 'command 401 not implemented' problems,
though probably not all of them.

There's definitely some bugs where rolling back past the start of an
interaction will remove all the images on the screen, probably
leading to situations where images can get out of sync.

Rollback can be completely disabled for a game by putting the key
'disable_rollback' in the rpgm_game_data file