Fix inability to save to map cache for certain pathnames
If the pathname contains non-ASCII characters there is some chance
that pygame_sdl2.image.save will fail to save to it. The problem
seems to get better if the filename is explicitly encoded as UTF-8
before being sent to pygame_sdl2, but there are probably some
environments where this is not the right thing to do.