Implement YEP_GridFreeDoodads plugin
Animated doodads only show the first frame

Doodads z-order is possibly supposed to have some interplay
with the normal tiles and sprites, but right now all the doodads
are on top of the tiles and below the sprites.

Doodads are baked into the map (in case there's a ton of them).
Since the number of doodads can change on a given map based on switch
values, the doodad switch conditions are part of the map file name.

Needed for Living With Mia Act 3.