rpgm_game_metadata can now contain 'map_zoom_rect' entries for better map display
If a zoom rect is specified, every time the player is transferred
to that map the user zoom level and scroll position will be set
to center the rect on the screen.

This is for The Artifact series where the creator added a lot of extra
blank space around the actual walkable area on the map. This area is
filled with invisible events that get used during scenes / animation
but never need to be seen by the player.