Augment 'eval_fancypants_value_statement' to be able to return literals
It will also return the literal result of a variable lookup statement
if that's all that's left in the string. Which is helpful if the variable
currently contains a string and you fant eval_fancypants_value_statement
to return that string.

for some of the stringy script eval in college-life