Fix renpy 7.x compatibility by showing 'maplayer' on top of 'master'
Renpy 7.x programs start with a 'scene black' instead of 'scene'
(transparent), which under the old configuration would cover the maplayer
(where practically everything is drawn) with black.

One solution is to run a 'scene' command after Renpy's 'scene black',
which would replace the background with a transparent one. But it might be
better to have black background anyway because it covers the transparent
checkerboard RenPy development background leading to a more authentic
production experience.

So instead 'maplayer' is now always on top of 'master'.

Changed the show/hide movie command to render them to 'maplayer' instead of
'master', otherwise they would show under the currently rendered picture.