Respect solid parallel events when calculating reachability
As needed for the-artifact series which often moves the player
to 'd-pad' shaped formations while showing a picture; the player
is intended to be trapped in such a way where hitting up-down-left-right
or the action button will trigger one of the identical events to the
up/down/left/right.

Sometimes these events are parallel, which should be solid as far
as reachability is concerned provided their 'priorityType' is set to 1.