Preserve list/dict references to better facilitate rollback
RenPy turns [] and {} into RevertableList and RevertableDict instances
respectively, and they're easier for RenPy to revert them when rolling
back if you empty them instead of clobbering the instance.

This means:
* `arr = []` becomes `del arr[:]`
* `dic = {}` becomes `dic.clear()`

i'm not actually sure if those statements are rollback friendly,
because the real fix in this diff was to not reset game_state.events
using a list comprehension every time the event queue burns down.

instead, the next auto-trigger event (if available) is just appeneded
to the empty `.events` list.

also was important to preserve objects references for `parallel_events`
and `move_routes` arrays by deleting then re-adding content rather
than re-assigning

this fixes issues with rolling back during 'dpad animations' in the artifact series