Fix non-player character direction changes to show correctly
Non-player direction changes were registered to page 'None' of the
event, which I think is appropriate because the character direction
sticks even if the page changes. But the event_sprite_data function
wasn't aware of that.

hopefully fixes character display in any number of places, notably
in timed sequences in the_artifact