Retool how page properties are saved a little bit to fix bugs
* direction overrides are saved per-page, but often carryover
when switching pages (the way the MV code treats it, the direction
only gets reset if the "originalDirection" differs in the old vs new page)

fixes the visible direction of the rival cheer alarm scene in the artifact part 3

* lazily clear out overrides whenever the active event page changes

fixes the display of cheer items on the bottom floor of the rival cheer HQ
from not resetting back to the original item when you put an item back

refactoring-wise, this introduces a bunch of methods and constants to try
to make dealing with these page overrides more ergonomic. is it? who
knows!