Add a 'Skip event' button to certain scenes
SkippableManager is capable of doing a bunch of different commands
as specified in the rpgm_game_data file to skip past an event,
including setting switches, self switches, variables, etc.

Completely replaces the old 'switch_triggers' system: events that
used that system are now just 'skippables' with 'auto: true' set.

As seen mostly in the-artifact part 3