Process triggered common events before any subsequent events
In the RPGMaker MV code, this is handled by creating a 'childInterpreter'
anytime a common event is invoked. All the commands in the
childInterpreter need to be processed before any parallel/normal events
outside of it resume. I think.

In this case, I'm just making an extra array of events called
'triggered_common_events' to supplement the normal 'events'
and 'parallel_events' arrays. Hopefully it works out to be about the same.

primarily created to fix issues with the mushroom code in the big
overworld of warlock-and-boobs: there's a bunch of common events stuff that
wasn't happening in the right order, which is one of the causes of
infinite loop looking stuff when you start fiddling around south of the
main forest