Fix diagonal movement in move routes to not crash after game has been reloaded
Weirdly the global symbol 'map' is replaced with the value of
(presumably) game_state.map after a game has been reloaded once. This is
bad, and could probably be fixed, but for now the hack is to reference
the builtin Python 'map' from __builtins__ to ensure the code doesn't
crash trying to use a GameMap object for functional programming

bonus: fix an 'event_through' that should have become 'page_property' in an earlier refactor

needed for any game with diagonal movement, probably peasant's quest?