Fix 'YEP_QuestJournal' support to track objective/reward visibility different than status
Previously, if a quest was marked as 'COMPLETE ALL OBJECTIVES', the
objectives would all have status set to 'completed' and show up in the
quest log.

Now, objectives, regardless of whether they are completed, failed, or
otherwise, only show up if there has been a 'SHOW OBJECTIVE' call
earlier

without this, peasant's quest currently crashes trying to render a
placeholder objective after a certain point. but it is also needed to
just not show the wrong objectives.