MV export script no longer squashes decorative events into tiles
Some games may need to keep these events around to operate on with a
script (e.g. dynamically changing or moving an image on the map)

one example is the cage opening at the end of urban demons

there's probably a performance penalty if a map has a LOT of these
events, because sprite events are more costly to render in rpgm2renpy
than tile images that get baked into the rpgmcache map. but that could
probably be worked around if it becomes an issue.